also separate specific ParquetFile tests + remove test_empty
